Add new CommandTriggerAction
