+++ src/MahApps.Metro/MahApps.Metro.Shared/MahApps.Metro.Shared.projitems
@@ -10,6 +10,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Actions\CloseTabItemAction.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Actions\CommandTriggerAction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Actions\SetFlyoutOpenAction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Behaviours\BindableResourceBehavior.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Behaviours\BorderlessWindowBehavior.cs" />
